Bugfix: Resolve source line link offset
`godoc` always subtracted 10 from the the line number which results
in the wrong line being linked to.
Copies the code from `godoc` to remove the single line to fix the
formatting issue.

Enables the source link hash to be configurable to support source
control platforms other than Github. The default value remains the
same as before but is now configurable for other platforms.

Regenerate examples to verify the only changes are due to fix applied
as part of this commit.